moved openapi.go functions into apps.go
